+++ examples/basic/src/bin/hello_no_macros.rs
@@ -1,4 +1,5 @@
-use moonbeam::{AsyncFnServer, Body, Request, Response, Server, Spawner, StatelessAsyncFnServer};
+use moonbeam::server::{AsyncFnServer, LifetimeDummy, StatelessAsyncFnServer};
+use moonbeam::{Body, Request, Response, Server, Spawner};
 
 // Note: we're using 'static here to make the example easier to read, but this could be an arbitrary
 // lifetime (e.g. 'a) that allows the server to hold a reference to objects created during
@@ -14,17 +15,19 @@ impl Server for HelloNoMacros {
 	}
 }
 
-async fn no_state<'req, 'exec>(
+async fn no_state<'exec, 'req>(
 	_request: Request<'req, 'req>,
 	_spawner: Spawner<'exec>,
+	_: LifetimeDummy<'exec, 'req>,
 ) -> Response<'req> {
-	Response::ok().with_body(format!("Hello no state"), Body::TEXT)
+	Response::ok().with_body("Hello no state".to_string(), Body::TEXT)
 }
 
-async fn with_state<'req, 'exec>(
+async fn with_state<'exec, 'req>(
 	_request: Request<'req, 'req>,
 	_spawner: Spawner<'exec>,
 	state: &'exec &str,
+	_: LifetimeDummy<'exec, 'req>,
 ) -> Response<'req> {
 	Response::ok().with_body(format!("Hello {}", state), Body::TEXT)
 }